requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-crispy-forms==2.0
django-intl-tel-input==0.3.1
sqlparse==0.4.3
